+++ lib/whois/record/parser/whois.bnnic.bn.rb
@@ -14,7 +14,7 @@ module Whois
   class Record
     class Parser
 
-      # Parser for the whois.bn server.
+      # Parser for the whois.bnnic.bn server.
       #
       # @note This parser is just a stub and provides only a few basic methods
       #   to check for domain availability and get domain status.
@@ -23,7 +23,7 @@ module Whois
       # @see Whois::Record::Parser::Example
       #   The Example parser for the list of all available methods.
       #
-      class WhoisBn < Base
+      class WhoisBnnicBn < Base
 
         property_supported :status do
           if available?
@@ -34,7 +34,7 @@ module Whois
         end
 
         property_supported :available? do
-          !!(content_for_scanner =~ /^No records matching .+ found/)
+          !!(content_for_scanner =~ /^Domain Not Found/)
         end
 
         property_supported :registered? do
@@ -43,23 +43,30 @@ module Whois
 
 
         property_supported :created_on do
-          if content_for_scanner =~ /Created:\s+(.+)\n/
+          if content_for_scanner =~ /Creation Date:\s+(.+)\n/
             Time.parse($1)
           end
         end
 
         property_supported :updated_on do
-          if content_for_scanner =~ /Last Updated:\s+(.+)\n/
+          if content_for_scanner =~ /Modified Date:\s+(.+)\n/
             Time.parse($1)
           end
         end
 
-        property_not_supported :expires_on
+        property_supported :expires_on do
+          if content_for_scanner =~ /Expiration Date:\s+(.+)\n/
+            Time.parse($1)
+          end
+        end
 
 
         property_supported :nameservers do
-          content_for_scanner.scan(/Name Server \d:\s+(.+)\n/).flatten.map do |name|
-            Record::Nameserver.new(:name => name)
+          if content_for_scanner =~ /Name Servers:\n((.+\n)+)\n/
+            $1.split("\n").map do |line|
+              name, ipv4 = line.strip.scan(/(.+) \((.+)\)/).flatten
+              Record::Nameserver.new(name: name.strip.downcase, ipv4: ipv4.split(",").first)
+            end
           end
         end